sql: expose logical times via mz_logical_timestamp builtin
Users often want to know the logical time at which their statement
executed. SQL makes this somewhat difficult, as queries cannot return
metadata outside of the requested result set. The traditional solution
is to expose a function that returns the transaction timestamp. In
PostgreSQL, `now()` and `transaction_timestamp()` both serve this
purpose. In CockroachDB, `cluster_logical_timestamp()` serves this
purpose. Users can then run queries like

  SELECT a, b, c, logical_timestamp()

where the logical timestamp is produced in a column alongside the
desired data.

Since our logical timestamps are u64s that are potentially unrelated to
wall clock time, this patch follows CockroachDB's lead in adding a new
function, `mz_logical_timestamp()`, which returns the logical timestamp
at which we've chosen to execute a query. Our `now()` function continues
to return wall clock time.

There is an additional wrinkle here in that a query like

    CREATE MATERIALIZED VIEW v AS SELECT mz_logical_timestamp()

that attempts to create a view over mz_logical_timestamp is very
challenging to support. What should the semantics of this view be?
Should it continually produce updates? And if so, at what frequency?

This patch punts on answering these questions by simply disallowing
calls to mz_logical_timestamp in view queries. For good measure, now is
also now disallowed in views, because its semantics are unclear for the
same reasons.

The implementation approach taken is to require an "evaluation
environment" when evaluating a scalar expression. This environment
at present contains the logical time and wall clock time, but will
eventually grow to contain other bits of state that can impact how
expressions are evaluated, e.g., the local timezone.

Fix #1199.