Pivot to new differential arrangement implementation.
The new differential arrangement implementation	allows us to decouple compaction of arrangements from the arrival of updates, which allows us to compact more aggressively. There are some general benefits to the new implementation, but it is also new and we might have some bugs to shake out (VMR is using it and it passes all tests there).

One important change is that differential now responds to an environment variable DIFFERENTIAL_EAGER_MERGE which can be set to an integer, and will prompt differential to feign that many phantom updates each time an arrangement operator is scheduled, if there is more than one distinct batch in the arrangement (informally: if there is merging work to do). This causes free cycles to be spent compacting arrangements, and gets us towards better compacted data without needing to feed actual updates at the arrangement.