pgwire: properly reject ssl/gssapi encryption requests
We were previously returning a fatal error when clients requested SSL or
GSSAPI encryption. The protocol allows for a softer method of returning
"SSL not supported" or "GSSAPI encryption not supported", which allows
clients to fall back to unencrypted connections. Switch to that softer
method, so that explicitly passing sslmode=disable or gssencmode=disable
is no longer required.