Add http /status page and time-of-build to root http page
This code is pretty ugly, but ideally it will allow us to make things nicer in the medium
term. It's not quite the absolute minimum I could do to get a
slightly-more-human-readable page up, but it seems like it should set us up to be able to
have different displays for different metric types as time goes on.