expr: fix bugs in ScalarExpr::non_null_requirements
This function can now be implemented in terms of
[Nullary | Unary | Binary | Variadic ]Func::propagates_nulls,
which avoids duplicating information and also fixes a buglet where
BinaryFunc::And was not propertly marked as not propagating nulls.

While I'm in here, fix a bug in ScalarExpr::If, which does *not*
necessarily propagate null (thanks, SQL) if the condition is null,
because the else condition will still be evaluated.