/status: fix metric aggregation/summing
Individual metric families contain an arbitrary number of individual measurments,
separated out by labels. This was a bit hidden by the fact that rust-prometheus exposes
`RepeatedField` properties as getters that are named like they return a single value,
although that might just be a knock-on effect of the proto schema.

This changes us to not just select the zeroeth element of the list.