peeker: upgrade to postgres v0.17
This is part of a patch series to standardize on postgres v0.17, so we
don't have two versions of the dependency floating around in the crate
graph.

This change set is unfortunately large because of changes to the Tokio
libraries that require e.g. the runtime context to bind to an address.