rust-toolchain: upgrade to Rust 1.40.0
Option::as_deref was stablized in Rust 1.40, so we can drop our version
of it, mz_as_deref, from ore::option::OptionExt.

Clippy now caches warnings, so we no longer need to clean before running
`cargo clippy`.

The other changes in this PR are to comply with Clippy's new lints,
which are mostly about avoiding unnecessary calls to `.clone()`.