materialized: avoid build script
The materialized crate currently uses a build script to determine the
current Git SHA and time so that those values can be embedded into the
binary. This is accepted best practice at the moment, but I think this
this practice is misguided, for several reasons:

  * Build scripts are only re-run if something in the materialized crate
    changes. It is perfectly possible--likely, even--to switch to a
    commit that only has changes in one of the materialized crate's
    dependencies, in which case the next build will *not* rerun the
    build script and will fail to noticy the updated commit SHA and
    build time, even though the materialized crate will be rebuilt.

  * Build scripts *are* rerun if something in the crate changes, which
    invalidates the crate library. This is painful when adjusting
    integration tests in the materialized/tests directory, where a small
    change to a test can result in 30s+ of recompilation of the library.

  * The `cargo:rerun-if-changed` directives are not sufficiently
    powerful to request a re-run whenever the crate is being recompiled.

Instead, we can (ab)use procedural macros to run code at compile time,
and shell out to `git` and `date` to get the Git SHA and current time,
respectively. Someone has already packaged this logic up into a crate
called `compile-time-run`, which works well and solves all of the above
deficiencies of build scripts.