ci: don't auto-update the Cargo.lock
If the Cargo.lock is out-of-date wrt to the Cargo.toml, CI shouldn't
pass the build, as that means that everyone who runs cargo locally on
that commit will generate diffs to Cargo.lock.

We can tell cargo not to auto-update the lockfile, and instead produce
an error, by passing --locked to every cargo invocation.