sql-parser: remove non-postgres dialects
We don't care about anything but the Postgres dialect. Attempting to
support all possible SQL dialects is not what we want to stake
Materialize's future on. Rip out support for dialects entirely, so we
can focus on the one dialect that matters: ours.

The code movement here is to consolidate the MySQL- and
PostgreSQL-specific tests that are still relevant in our new, merged
Materialize dialect into the main sqlparser test file.