*: localize clippy exceptions
Module-level clippy exceptions are almost universally too broad. Either
the lint is bad, and it should be ignored globally, or the lint is
generally good, but there is one specific place where it triggers a
false positive. The exception is not usually intrinsic to a specific
module, but rather a specific function or block; localize some clippy
lints accordingly.