Collapse RowPacker and PackableRow into one type
Rather than keep a long-lived buffer, allocate a large buffer upfront
and shrink it to produce the row.

The performance impact of this is hard to assess - it's very sensitive
to the state of the global allocator so eg changing the order of
microbenchmarks can heavily affect the results :(

But logically the cost should be a single extra free per Row created.

This patch also incidentally fixes a use-after-free in RowArena.