expr: allow functions to take typed data as input
Various functions want to take typed Rust data as plan-time input, in
addition to the zero, one, two, or variadic datums the functions take at
runtime. For example, BinaryFunc::FloorDecimal needs to know the scale
of the decimal at runtime, and ScalarExpr::MatchCachedRegex passes a
compiled regex::Regex to the match_regex function.

There are two approaches to this that have been taken so far. The first
is to stuff the input into a datum, like BinaryFunc::FloorDecimal does.
FloorDecimal is really a unary function--it naturally would only take
one datum as input--but we use a binary function so we can stuff the
scale into the second parameter. The second approach is what
ScalarExpr::MatchCachedRegex does, i.e., a dedicated ScalarExpr variant.
This approach is particularly ugly, as every ScalarExpr transformation
now needs to handle MatchCachedRegex explicitly.

This pain is all caused because we need to construct a uniform function
ponter type for all {nullary, unary, binary, variadic} functions.
Consider UnaryFunc, for example, which presently exposes this API:

    fn func(self) -> for<'a> fn(Datum<'a>, &EvalEnv, &'a RowArena) -> Datum<'a>

Every UnaryFunc must therefore take exactly one datum, an evaluation
environment, and a row buffer. There is no provision for a function like
UnaryFunc::FloorDecimal, which takes an additional u8 scale parameter.

This commit adjusts UnaryFunc to instead expose this API:

    fn eval<'a>(
        &'a self,
        a: Datum<'a>,
        _env: &'a EvalEnv,
        temp_storage: &'a RowArena,
    ) -> Datum<'a>

Now the caller has no insight into how the individual functions are
implemented. floor_float32 has a very simple signature

    fn floor_float32<'a>(a: Datum<'a>) -> Datum<'a>

while floor_decimal can take the extra parameter it desires

    fn floor_decimal<'a>(a: Datum<'a>, scale: u8) -> Datum<'a>

and only functions that use temp_storage need to take it as a parameter:

    fn cast_int32_to_string<'a>(
        a: Datum<'a>,
        temp_storage: &'a RowArena,
    ) -> Datum<'a>

Analogous changes are made to the other function types, including the
relational functions, to keep everything symmetric.
BinaryFunc::FloorDecimal, BinaryFunc::CeilDecimal, and
BinaryFunc::CastDecimalToString all become UnaryFuncs, as they naturally
should be. ScalarExpr::MatchCachedRegex is replaced with
UnaryFunc::MatchRegex, which removes a nice amount of special cases from
the optimizer.

This also paves the way for a UnaryFunc::CastStringToDecimal function,
which can't currently be expressed as UnaryFunc::output_type has no way
to derive the output type for UnaryFunc::CastStringToDecimal without
this patch (as how would it figure out what scale to report?).