Remove redundant join inputs.
In some cases, often due to subqueries, we have joins that join the same relation multiple times, using a key of the relation. In these cases, all but one instance can be removed, with references to the other relations set to the first, and a projection wrapping the whole thing to set up columns appropriately. Perhaps there are other transformations we should apply to logically remove the columns, but let's see what this looks like.