dataflow: poll tailed files every 100ms on macOS
We expect file sources to be used primarily for demo purposes, so it
makes sense to have the default be tuned for a good demo experience, and
not being able to support massive amounts of file sources. (The default
before was 30s, which made Materialize seem rather unimpressive.) Plus
with enough engineering effort we could use something like kqueue on
macOS instead, which provides precise notifications of when a file has
been written.