testdrive: Correct proto serialization
This adds some infrastructure to verify that we're actually writing the correct bytes to
the wire, because apparently the old code was doing the wrong thing, and some of the
other methods I tried also failed. The extra `validator` call failed with the old code,
and succeeds correctly now.

The testdrive test is now verifying various parts of the ingested json structure, instead
of the full one. (Issue #1533)

With commit I think we can say that this PR Fixes #1398