protobuf: Just support passing descriptors as base64
This simplifies the compile protoc command such that it is no longer an action, it just
tries to encode the proto descriptor first as a file then as base64, which we pass as a
literal b64 value in the schema creation sql statement.

We then don't need protoc on our build servers, we continue to check the compiled
descriptor files into git. An obvious extension to this is to just have some well-known
descriptors that we embed in the test-drive binary, but that requires significantly more
code churn on every descriptor addition, which doesn't seem worth it.