pgwire: fix memory leak if client disconnects during blocked tail
If a client issues a `TAIL never_changing_source` statement and
then disconnects, we would fail to notice the client disconnect, and
block a Tokio task forever on waiting for that never changing source to
change. This presented as test_file_sources wedging if it panicked in
the wrong place, but in production was likely to cause a memory leak, as
these forever-wedged connections could pile up.

This commit fixes the issue by introducing an infrequent (1s) keepalive
packet. If a tailed source doesn't change for a full second, we send the
client a message that indicates no new data; if that message fails to
send, the client has gone away, and we can clean up that connection on
our end. This solution is somewhat distasteful, but no better solution
presented itself; justification is provided in a code comment within the
patch..