dc.sh: add a status command
Because of the hackering of parsing cli output I had to change an internal variable name
and a visible container name.

It currently just tries to show the exposed ports and running containers:

    $  dcsh status
    CONTAINER                            PORT
    =========                            ====
    materialized                         6875
    metabase                             3030
    ..snip..
    prometheus_sql_exporter
    prometheus_sql_exporter_mysql_tpcch
    control-center                       9021

But at least it shows something, and I often just want to check what's running and
remember what ports they're listening on.