sql: use correct column names for complicated aggregates
Complicated aggregates like avg and stddev are decomposed into their
constituent parts by the SQL transformation pipeline. This pipeline runs
early enough that it impacts the column names chosen for these aggregate
functions; the avg function was resulting in no column name, while the
stddev function was resulting in the column name `sqrt`. Fix this by
teaching the transformation pipeline to preserve the name of the
original function when necessary.