catalog: transactionalize api
Rework the catalog API so that updates happen transactionally. This is
necessary for DROP {DATABASE|SCHEMA}, where dropping a database or
schema requires dropping all the items contained within atomically. It
also, incidentally, fixes a bug in DROP VIEW ... CASCADE, where removal
of the cascaded objects would not be persisted atomically.