Improve robustness of Reduce implementation.
The implementation of Reduce is improved in three ways. First, each reduction is performed independently, to avoid maintaining a cross-product of values. Second, distinct aggregates are applied in a distinct operator, which can reduce the churn. Third, hierarchical aggregations like min and max are applied hierarchically to minimize the update latency.