Be more careful about tracking columns for outer relations
When there are both multiple subqueries in one WHERE and some of them
are nested, the working columns for the earlier subqueries can get
included in the outer relation for the later subqueries. We have to
keep track of these to avoid off-by-one errors in references to the outer.

Fixes https://github.com/MaterializeInc/materialize/issues/1708