Clear render cache between views.
Each view uses various local identifiers in building its collections and arrangements, and there is no guarantee that they will be unique across views. This causes problems with false hits and panics if we attempt to rebind any definitions. To address this, we clear out from the context that is not a `Get` of a global identifier. Each view builds a collection (and perhaps arrangements) bound to its global identifier, and each index build instruction is around a global get for the view (we do not allow you to index a non-view query).