chbench: add docker build script and fix docker build/compose
This changes the chbenchmark docker repo to use the same standard we use for other
in-tree build projects, where they are required to be manually built, but also expose a
`:local` tag for development within docker-compose.

One thing worth considering here is that it will change the meaning of the
`materialize/chbenchmark` repo in docker-hub, previously it did not include all our
configuration for odbc, but now it will.