catalog,coord,dataflow,sql: split view/source/sink/index types
The catalog, coord, dataflow, and sql crates previously all shared the
same view, source, sink, and index types in dataflow_types. But these
crates all had slightly different opinions on what should live in the
types. The sql crate, for example, does not want to care about eval
envs, nor does it want to duplicate type information in the catalog; the
dataflow layer *needs* duplicated type information, because it doesn't
have access to a catalog.

Start untangling the mess by giving catalog, dataflow, and sql their own
view/source/sink/index types. The coordinator converts between the types
as necessary.

This commit also introduces the concept of a CatalogItemSerializer, so
that the catalog is agnostic with regards to the serialization strategy
for items themselves. This will allow us to easily use SQL to serialize
catalog items in a future commit.