coord,sql: use normalized SQL to describe catalog objects
Storing serialized Rust objects into the catalog was convenient, but had
poor backwards-compatibility characteristics. Any change to our internal
data structures caused stored catalogs to be unreadable.

SQL is a far more stable API to Materialize. We won't break
backwards-compatibility in the SQL API lightly. So use SQL to describe
the items in the catalog instead.

This commit introduces the necessary serialization and deserialization
logic, as well as the normalization rules that make it safe to store a
`CREATE { SOURCE | VIEW | INDEX | SINK }` in the catalog (i.e., by
quoting identifiers, resolving names, etc.). The new SQL serializer is
not yet hooked up, as there is additional prepatory work required to
build the create SQL for sources and auto-created indexes.