coord: properly persist auto-created primary keys on ambiguous columns
CREATE MATERIALIZED VIEW can create an index on columns in a view whose
names are ambiguous, either because the column has no name or because
two columns have the same name. We were properly detecting the former
case, but not the latter. In either case the column reference needs to
use a column index instead of the column name.

This commit fixes the issue and adds a test case.

Fix #1900.